fix: updated route structuring
