+++ app/(proctected)/events/[id]/page.tsx
