+++ app/(proctected)/events/filter-bar.tsx
