+++ app/(proctected)/layout.tsx
@@ -1,0 +1,22 @@
+import type { Metadata } from "next";
+import { LayoutSetter } from "@/components/layouts/layout-setter";
+
+
+export const metadata: Metadata = {
+  title: "Genekriti",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <body>
+        <LayoutSetter>{children}</LayoutSetter>
+      </body>
+    </html>
+  );
+}
