+++ app/(proctected)/market-place/search-bar.tsx
@@ -1,21 +1,22 @@
 "use client";
 import { XCircle } from "lucide-react";
-import { Button } from "../../components/ui/button";
-import { Input } from "../../components/ui/input";
+import { Button } from "../../../components/ui/button";
+import { Input } from "../../../components/ui/input";
 import { FilterBar } from "./filter-bar";
 import { useState } from "react";
 import { useSearchParams } from "next/navigation";
 
 export const SearchBar = () => {
-    const params = useSearchParams();
-    const [search, setSearch] = useState(params.get("search") || "");
+    const [updatedTags, setUpdatedTags] = useState(["All"]);
+    const [search, setSearch] = useState(useSearchParams().get("search") || "");
+
     return (
         <div className="sticky top-[96px] left-0 bg-accent flex flex-col items-center gap-3 pt-4 pb-3">
             <div className="flex w-full items-center">
                 <form
                     onSubmit={(e) => {
                         e.preventDefault();
-                        window.location.href = `/events?search=${search}`;
+                        window.location.href = `/market-place?search=${search}`;
                     }}
                     className="w-full"
                 >
@@ -31,13 +32,16 @@ export const SearchBar = () => {
                 <Button
                     variant="ghost"
                     onClick={() => {
-                        window.location.href = "/events";
+                        window.location.href = "/market-place";
                     }}
                 >
                     <XCircle stroke="#800000" />
                 </Button>
             </div>
-            <FilterBar />
+            <FilterBar
+                updatedTags={updatedTags}
+                setUpdatedTags={setUpdatedTags}
+            />
         </div>
     );
 };
